Vue implementation
I added Vue as a frontend. The Flask returns and gets json, you can upload new music and there are separate pages for every song where Vue gets the columns of every song which includes name, creator, genre and the audio file name that is stored in a separate folder.